add class for totalStaked detail

+++ libs/OMMStruct/src/main/java/finance/omm/libs/structs/TotalStaked.java
@@ -3,6 +3,6 @@ package finance.omm.libs.structs;
 import java.math.BigInteger;
 
 public class TotalStaked {
-    public Integer decimals;
+    public BigInteger decimals;
     public BigInteger totalStaked;
 }
